Added User Auth, Manage Gigs and Relationships

+++ app/Http/Controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\Listing;
 
+use App\Models\User;
+
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule as ValidationRule;
 
@@ -51,6 +53,9 @@ class ListingController extends Controller
         }
 
 
+        $formFields['user_id'] = auth()->id();
+
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing Created!');
@@ -80,6 +85,13 @@ class ListingController extends Controller
 
 
     public function update(Request $request, Listing $listing){
+
+        //Make sure logged in user is owner
+
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -104,7 +116,18 @@ class ListingController extends Controller
 
     }
 
+    //Show manage listing
+
+    public function manage(){
+        return view('listings.manage',['listings' => auth()->user()->listings()->get()]);
+    }
+
     public function destroy(Listing $listing){
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
+
+
         $listing->delete();
         return redirect('/')->with('message','Listing Deleted Successfully!');
