+++ routes/web.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\ListingController;
+use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Models\Listing;
@@ -41,34 +42,61 @@ Route::get('/', [ListingController::class, 'index']);
 
 //Show create form
 
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 
 
 //Store Listing Data
 
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 //Show Edit Form
 
-Route::get('/listings/{listing}/edit', [ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class,'edit'])->middleware('auth');
+
+
+
+
+//Show the register page
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
+
+
+//Create new User
+Route::post('/users', [UserController::class, 'store']);
+
+
+//logout user
+
+Route::post('/logout', [UserController::class,'logout'])->middleware('auth');
+
+
+//Show Login Form
+
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
+
+//Log in User
+
+Route::post('users/authenticate',[UserController::class,'authenticate']);
+
+
+
+//Show Manage Gigs
+
+Route::get('/listings/manage',[ListingController::class,'manage'])->middleware('auth');
+
+
 
 //Update Listing
 
-//Traversity
-Route::put('/listings/{listing}', [ListingController::class,'update']);
 
+//Traversity
+Route::put('/listings/{listing}', [ListingController::class,'update'])->middleware('auth');
 
 //Mr Haree Update Listing
 //Route::put('/listings/{listing}', 'ListingController@update')->name('listingUpdate');
 
 
-
-
-
-
-
 //Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class,'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class,'destroy'])->middleware('auth');
 
 
 //Single Listing
